Se agregan los siguientes cambios: - Autenticacion simple de Django y autenticacion por JSONWebToken - Se agrega el endpoint para borrar un ingrediente de una pizza creada - Excepcion: Si una pizza tiene un ingrediente, este no se podra borrar - Se protegen todos los endpoints excepto los de creacion de usuarios

+++ app/product/views.py
@@ -8,17 +8,23 @@ from product.serializers import (
 from rest_framework.response import Response
 from rest_framework import status
 from product.utils import resource_checker
+from rest_framework.permissions import IsAuthenticated
 
 
 ############################## Endpoints of Pizzas #################################
 
 
 class PizzasAPIView(APIView):
-  
+    permission_classes = (IsAuthenticated,)
     def get(self, request, format=None):
-        pizzas = Pizza.objects.all().filter(is_active=True)
-        serializer = PizzaSerializer(pizzas, many=True)
-        return Response(serializer.data)
+        if request.user.is_superuser:
+            pizzas = Pizza.objects.all()
+            serializer = PizzaSerializer(pizzas, many=True)
+            return Response(serializer.data)
+        else:
+            pizzas = Pizza.objects.all().filter(is_active=True)
+            serializer = PizzaSerializer(pizzas, many=True)
+            return Response(serializer.data)
       
     def post(self, request, format=None):
         serializer = PizzaSerializer(data=request.data)
@@ -29,7 +35,7 @@ class PizzasAPIView(APIView):
       
 
 class PizzaDetailAPIView(APIView):
-
+    permission_classes = (IsAuthenticated,)
     @resource_checker(Pizza)
     def get(self, request, pk, format=None):
         pizza = Pizza.objects.get(pk=pk)
@@ -50,13 +56,25 @@ class PizzaDetailAPIView(APIView):
         pizza = Pizza.objects.filter(id=pk).first()
         pizza.delete()
         return Response({"message": f"Pizza '{pizza}' deleted succesfully"}, status=status.HTTP_204_NO_CONTENT)
+      
+      
+class DeleteIngredientPizza(APIView):
+    """Delete an ingredient of a pizza"""
+    permission_classes = (IsAuthenticated,)
+    def delete(self, request, pizza_id, ingredient_id, format=None):
+        pizza = Pizza.objects.get(id=pizza_id)
+        ingredient = Ingredient.objects.get(id=ingredient_id)
+        pizza.ingredients.remove(ingredient)
+        return Response({
+          "message": f"Se ha removido el ingrediente {ingredient} de la pizza {pizza}"
+        })
 
 
 ############################# Endpoints of Ingredients #############################
 
 
 class IngredientsAPIView(APIView):
-
+    permission_classes = (IsAuthenticated,)
     def get(self, request, format=None):
         ingredients = Ingredient.objects.all()
         serializer = IngredientSerializer(ingredients, many=True)
@@ -71,7 +89,7 @@ class IngredientsAPIView(APIView):
       
       
 class IngredientDetailAPIView(APIView):
-
+    permission_classes = (IsAuthenticated,)
     @resource_checker(Ingredient)
     def get(self, request, pk, format=None):
         ingredient = Ingredient.objects.get(pk=pk)
@@ -90,15 +108,20 @@ class IngredientDetailAPIView(APIView):
     @resource_checker(Ingredient)
     def delete(self, request, pk, format=None):
         ingredient = Ingredient.objects.filter(id=pk).first()
-        ingredient.delete()
-        return Response({"message": f"Ingredient '{ingredient}' deleted succesfully"}, status=status.HTTP_204_NO_CONTENT)
+        pizza_with_that_ingredient = Pizza.objects.filter(ingredients=ingredient).exists()
+        # Si existe una pizza asociada con ese ingrediente no se podrá eliminar la misma.
+        if pizza_with_that_ingredient:
+            return Response({"message": "There is a pizza with that ingredient, it cannot be deleted"})
+        else:
+            ingredient.delete()
+            return Response({"message": f"Ingredient '{ingredient}' deleted succesfully"}, status=status.HTTP_204_NO_CONTENT)
 
 
 ############################# Endpoints of Categories #############################
 
 
 class CategoriesAPIView(APIView):
-  
+    permission_classes = (IsAuthenticated,)
     def get(self, request, format=None):
         categories = Category.objects.all()
         serializer = CategorySerializer(categories, many=True)
@@ -113,7 +136,7 @@ class CategoriesAPIView(APIView):
       
       
 class CategoryDetailAPIView(APIView):
-    
+    permission_classes = (IsAuthenticated,)
     @resource_checker(Category)
     def get(self, request, pk, format=None):
         category = Category.objects.get(pk=pk)
